+++ src/main/java/com/b1n_ry/yigd/config/YigdConfig.java
@@ -254,11 +254,11 @@ public class YigdConfig implements ConfigData {
         @ConfigEntry.Gui.Tooltip
         @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)
         public DeathEffectConfig claimRuleOverride = DeathEffectConfig.CREATE_GRAVE;
-//
-//        @ConfigEntry.Gui.Tooltip
-//        public boolean prioritiseTheGraveyardGraves = false;
-//        @ConfigEntry.Gui.Tooltip
-//        public int graveyardSearchRadius = 10;
+
+        @ConfigEntry.Gui.Tooltip
+        public boolean prioritiseTheGraveyardGraves = false;
+        @ConfigEntry.Gui.Tooltip
+        public int graveyardSearchRadius = 10;
 
         public boolean keepAllTrinkets = false;
     }